Refactoring project overview page, started working on copy FFT values dialog
